refactor: move Datum, Rotation, Position to commons
