+++ src/chart_types/xy_chart/annotations/annotation_marker.test.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
-
-import { AnnotationDomainTypes, AnnotationSpec, Position, Rotation, AnnotationTypes } from '../utils/specs';
+import { AnnotationDomainTypes, AnnotationSpec, AnnotationTypes } from '../utils/specs';
+import { Position, Rotation } from '../../../utils/commons';
 import { DEFAULT_ANNOTATION_LINE_STYLE } from '../../../utils/themes/theme';
 import { Dimensions } from '../../../utils/dimensions';
 import { GroupId } from '../../../utils/ids';
