+++ src/chart_types/xy_chart/state/chart_state.test.ts
@@ -4,15 +4,14 @@ import {
   AnnotationTypes,
   AxisSpec,
   BarSeriesSpec,
-  Position,
   RectAnnotationSpec,
   SeriesTypes,
 } from '../utils/specs';
+import { Position } from '../../../utils/commons';
 import { TooltipType, TooltipValue } from '../utils/interactions';
 import { ScaleBand } from '../../../utils/scales/scale_band';
 import { ScaleContinuous } from '../../../utils/scales/scale_continuous';
 import { ScaleType } from '../../../utils/scales/scales';
-// import { ChartStore } from './chart_state';
 import { IndexedGeometry, GeometryValue, BandedAccessorType } from '../../../utils/geometry';
 import { AxisTicksDimensions, isDuplicateAxis } from '../utils/axis_utils';
 import { AxisId } from '../../../utils/ids';
