+++ src/chart_types/xy_chart/state/selectors/get_tooltip_values_highlighted_geoms.ts
@@ -8,7 +8,8 @@ import { getComputedScalesSelector } from './get_computed_scales';
 import { getElementAtCursorPositionSelector } from './get_elements_at_cursor_pos';
 import { IndexedGeometry } from '../../../../utils/geometry';
 import { getSeriesSpecsSelector, getAxisSpecsSelector } from './get_specs';
-import { BasicSeriesSpec, AxisSpec, Rotation } from '../../utils/specs';
+import { BasicSeriesSpec, AxisSpec } from '../../utils/specs';
+import { Rotation } from '../../../../utils/commons';
 import { getTooltipTypeSelector } from './get_tooltip_type';
 import { formatTooltip } from '../../tooltip/tooltip';
 import { getTooltipHeaderFormatterSelector } from './get_tooltip_header_formatter';
