+++ src/chart_types/xy_chart/utils/axis_utils.ts
@@ -6,13 +6,12 @@ import {
   CompleteBoundedDomain,
   DomainRange,
   LowerBoundedDomain,
-  Position,
-  Rotation,
   TickFormatter,
   UpperBoundedDomain,
   AxisStyle,
   TickFormatterOptions,
 } from './specs';
+import { Position, Rotation } from '../../../utils/commons';
 import { AxisConfig, Theme } from '../../../utils/themes/theme';
 import { Dimensions, Margins } from '../../../utils/dimensions';
 import { AxisId } from '../../../utils/ids';
