+++ src/chart_types/xy_chart/utils/specs.ts
@@ -9,7 +9,7 @@ import {
   RectAnnotationStyle,
 } from '../../../utils/themes/theme';
 import { Accessor, AccessorFormat } from '../../../utils/accessor';
-import { RecursivePartial } from '../../../utils/commons';
+import { RecursivePartial, Color, Position, Datum } from '../../../utils/commons';
 import { AxisId, GroupId } from '../../../utils/ids';
 import { ScaleContinuousType, ScaleType } from '../../../utils/scales/scales';
 import { CurveType } from '../../../utils/curves';
@@ -17,11 +17,7 @@ import { RawDataSeriesDatum, SeriesIdentifier } from './series';
 import { AnnotationTooltipFormatter } from '../annotations/annotation_utils';
 import { Spec, SpecTypes } from '../../..';
 import { ChartTypes } from '../..';
-import { Datum } from '../../../utils/domain';
 
-export type Rotation = 0 | 90 | -90 | 180;
-export type Rendering = 'canvas' | 'svg';
-export type Color = string;
 export type BarStyleOverride = RecursivePartial<BarSeriesStyle> | Color | null;
 export type PointStyleOverride = RecursivePartial<PointStyle> | Color | null;
 
@@ -483,20 +479,6 @@ export interface AxisStyle {
   /** Specifies the amount of padding on the tick label bounding box */
   tickLabelPadding?: number;
 }
-
-/**
- * The position of the axis relative to the chart.
- * A left or right positioned axis is a vertical axis.
- * A top or bottom positioned axis is an horizontal axis.
- */
-export const Position = Object.freeze({
-  Top: 'top' as 'top',
-  Bottom: 'bottom' as 'bottom',
-  Left: 'left' as 'left',
-  Right: 'right' as 'right',
-});
-
-export type Position = $Values<typeof Position>;
 
 export const AnnotationTypes = Object.freeze({
   Line: 'line' as 'line',
