+++ src/mocks/specs/specs.ts
@@ -1,4 +1,4 @@
-import { mergePartial } from '../../utils/commons';
+import { mergePartial, Position } from '../../utils/commons';
 import {
   SeriesSpecs,
   DEFAULT_GLOBAL_ID,
@@ -9,7 +9,6 @@ import {
   LineSeriesSpec,
   BasicSeriesSpec,
   SeriesTypes,
-  Position,
 } from '../../chart_types/xy_chart/utils/specs';
 import { getSpecId, getGroupId } from '../../utils/ids';
 import { ScaleType } from '../../utils/scales/scales';
