+++ src/state/selectors/get_chart_rotation.ts
@@ -1,7 +1,7 @@
 import createCachedSelector from 're-reselect';
 import { getSettingsSpecSelector } from './get_settings_specs';
 
-import { Rotation } from '../../chart_types/xy_chart/utils/specs';
+import { Rotation } from '../../utils/commons';
 import { getChartIdSelector } from './get_chart_id';
 
 export const getChartRotationSelector = createCachedSelector(
