+++ stories/annotations.tsx
@@ -21,7 +21,7 @@ import { Icon } from '../src/components/icons/icon';
 import { KIBANA_METRICS } from '../src/utils/data_samples/test_dataset_kibana';
 import { getChartRotationKnob, arrayKnobs } from './common';
 import { BandedAccessorType } from '../src/utils/geometry';
-import { Position } from '../src/chart_types/xy_chart/utils/specs';
+import { Position } from '../src/utils/commons';
 
 const dateFormatter = timeFormatter('HH:mm:ss');
